+++ src/include/oeis_manager.hpp
@@ -66,7 +66,8 @@ class OeisManager {
                    const std::string& submitted_by) const;
 
   void alert(Program p, size_t id, const std::string& prefix,
-             const std::string& color, const std::string& submitted_by) const;
+             const std::string& color, const std::string& submitted_by,
+             bool tweet) const;
 
   friend class OeisMaintenance;